+++ crypto/src/main/java/br/com/pedroxsqueiroz/bt/crypto/dtos/Wallet.java
@@ -2,6 +2,7 @@ package br.com.pedroxsqueiroz.bt.crypto.dtos;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
@@ -9,14 +10,14 @@ import java.util.Objects;
 @Data
 public class Wallet {
 
-    Map< StockType, Double > stocksToAmmounts;
+    Map< StockType, BigDecimal> stocksToAmmounts;
 
-    public void addAmmountToStock( StockType stock, Double ammount )
+    public void addAmmountToStock( StockType stock, BigDecimal ammount )
     {
 
         if(Objects.isNull(this.stocksToAmmounts))
         {
-            this.stocksToAmmounts = new HashMap<StockType, Double>();
+            this.stocksToAmmounts = new HashMap<StockType, BigDecimal>();
         }
 
         this.stocksToAmmounts.put( stock, ammount );
