+++ crypto/src/main/java/br/com/pedroxsqueiroz/bt/crypto/services/ExitAmmountGetter.java
@@ -5,9 +5,11 @@ import br.com.pedroxsqueiroz.bt.crypto.utils.config_tools.AnnotadedFieldsConfigu
 import br.com.pedroxsqueiroz.bt.crypto.utils.config_tools.Configurable;
 import lombok.experimental.Delegate;
 
+import java.math.BigDecimal;
+
 public abstract class ExitAmmountGetter extends Configurable {
 
-    abstract public Double get(TradePosition openTrade);
+    abstract public BigDecimal get(TradePosition openTrade);
 
     @Delegate(types = Configurable.class )
     private AnnotadedFieldsConfigurer configurer = new AnnotadedFieldsConfigurer(this);
